movie details head refactored, code splitted, lazy load on movie collection and screens
