+++ src/components/MovieDetailsBlock/MovieDetailsMedia/MovieDetailsMedia.tsx
@@ -1,14 +1,35 @@
 import React from "react";
-import { Layout } from "antd";
-
+import { Layout, Skeleton } from "antd";
 import MovieDetailsCast from "./MovieDetailsCast/MovieDetailsCast";
 import MovieDetailsScreens from "./MovieDetailsScreens/MovieDetailsScreens";
-import MovieDetailsCollection from "./MovieDetailsCollection/MovieDetailsCollection";
-import MovieDetailsRecs from "./MovieDetailsRecs/MovieDetailsRecs";
 import SiderMedia from "./SiderMedia/SiderMedia";
+import dynamic from "next/dynamic";
+import { useInViewport } from "ahooks";
 import type { Collection } from "@/redux/api/movies/types/MovieDetailsType";
 
 import styles from "./MovieDetailsMedia.module.scss";
+
+const MovieDetailsCollection = dynamic(
+  () => import("./MovieDetailsCollection/MovieDetailsCollection"),
+  {
+    loading: () => (
+      <div className={styles.skeletonLoader}>
+        <Skeleton active paragraph={{ rows: 5 }} />
+      </div>
+    ),
+  }
+);
+
+const MovieDetailsRecs = dynamic(
+  () => import("./MovieDetailsRecs/MovieDetailsRecs"),
+  {
+    loading: () => (
+      <div className={styles.skeletonLoader}>
+        <Skeleton active paragraph={{ rows: 5 }} />
+      </div>
+    ),
+  }
+);
 
 interface MovieDetailsMediaProps {
   id: number;
@@ -35,17 +56,28 @@ const MovieDetailsMedia: React.FC<MovieDetailsMediaProps> = ({
   budget,
   revenue,
 }) => {
+  const collectionRef = React.useRef<HTMLDivElement | null>(null);
+  const [isCollectionInViewport] = useInViewport(collectionRef);
+  const [isCollectionLoaded, setIsCollectionLoaded] = React.useState(false);
   const { Sider } = Layout;
+
+  React.useEffect(() => {
+    if (isCollectionInViewport && !isCollectionLoaded) {
+      setIsCollectionLoaded(true);
+    }
+  }, [collectionData, isCollectionInViewport, isCollectionLoaded]);
+
   return (
     <div className={styles.media}>
       <div className="app-container content-with-aside">
         <div className="media-content">
           <MovieDetailsCast id={id} />
           <MovieDetailsScreens id={id} />
-          {collectionData && (
+          <div className={styles.observer} ref={collectionRef}></div>
+          {isCollectionLoaded && collectionData && (
             <MovieDetailsCollection collectionData={collectionData} />
           )}
-          <MovieDetailsRecs id={id} />
+          {isCollectionLoaded && <MovieDetailsRecs id={id} />}
         </div>
         <Sider theme="light" className={styles.sider}>
           <SiderMedia
