+++ src/components/MovieDetailsBlock/MovieDetailsMedia/MovieDetailsRecs/MovieDetailsRecs.module.scss
@@ -1,0 +1,5 @@
+@import '@/styles/variables';
+
+.title {
+  color: $primaryColorContrast;
+}
