+++ src/redux/api/trending/slice.tsx
@@ -14,5 +14,6 @@ export const trendingApi = baseApi.injectEndpoints({
 });
 
 export const {
-  useGetTrendingAllQuery
+  useGetTrendingAllQuery,
+  useLazyGetTrendingAllQuery
 } = trendingApi;
